+++ example/lib/data/models/character_response_model.dart
@@ -1,5 +1,4 @@
 import 'package:json_annotation/json_annotation.dart';
-
 part 'character_response_model.g.dart';
 
 @JsonSerializable()
@@ -18,7 +17,8 @@ class Character {
     required this.image,
   });
 
-  factory Character.fromJson(Map<String, dynamic> json) => _$CharacterFromJson(json);
+  factory Character.fromJson(Map<String, dynamic> json) =>
+      _$CharacterFromJson(json);
   Map<String, dynamic> toJson() => _$CharacterToJson(this);
 }
 
@@ -32,7 +32,8 @@ class CharacterResponseModel {
     required this.info,
   });
 
-  factory CharacterResponseModel.fromJson(Map<String, dynamic> json) => _$CharacterResponseModelFromJson(json);
+  factory CharacterResponseModel.fromJson(Map<String, dynamic> json) =>
+      _$CharacterResponseModelFromJson(json);
   Map<String, dynamic> toJson() => _$CharacterResponseModelToJson(this);
 }
 
@@ -46,6 +47,7 @@ class InfoModel {
     this.next,
   });
 
-  factory InfoModel.fromJson(Map<String, dynamic> json) => _$InfoModelFromJson(json);
+  factory InfoModel.fromJson(Map<String, dynamic> json) =>
+      _$InfoModelFromJson(json);
   Map<String, dynamic> toJson() => _$InfoModelToJson(this);
 }