+++ lib/remote_data_kit.dart
@@ -1,17 +1,11 @@
 library;
 
+// Excepciones
 export 'src/exceptions/network_exceptions.dart';
-export 'src/mappers/dio_exception_mapper.dart';
-export 'src/client/api_client_builder.dart';
-export 'src/models/paginated_response.dart';
-export 'src/constants/network_constants.dart';
 
-export 'package:dio/dio.dart'
-    show
-        Dio,
-        BaseOptions,
-        Interceptor,
-        DioException,
-        Options,
-        RequestOptions,
-        ResponseType;
+// Modelos públicos
+export 'src/models/api_response.dart';
+export 'src/models/paginated_response.dart';
+
+// Fachada — punto de entrada único
+export 'src/facade/remote_data_kit_facade.dart';