+++ lib/src/client/api_client_builder.dart
@@ -1,25 +1,20 @@
 import 'package:dio/dio.dart';
 
-import '../constants/network_constants.dart';
-
 class ApiClientBuilder {
   static Dio build({
-    String? baseUrl,
-    int connectTimeout = NetworkConstants.defaultConnectTimeout,
-    int receiveTimeout = NetworkConstants.defaultReceiveTimeout,
-    List<Interceptor>? interceptors,
+    required String baseUrl,
+    required int connectTimeout,
+    required int receiveTimeout,
+    Map<String, String>? headers,
   }) {
     final dio = Dio(
       BaseOptions(
-        baseUrl: baseUrl ?? '',
+        baseUrl: baseUrl,
         connectTimeout: Duration(milliseconds: connectTimeout),
         receiveTimeout: Duration(milliseconds: receiveTimeout),
+        headers: headers,
       ),
     );
-
-    if (interceptors != null) {
-      dio.interceptors.addAll(interceptors);
-    }
 
     return dio;
   }